+++ tests/test_build.py
@@ -330,12 +330,12 @@ class TestBuildRss(BaseTest):
                         len(feed_item.description), 150, feed_item.title
                     )
 
-    def test_simple_build_lang_with_territory(self):
+    def test_simple_build_locale_with_territory(self):
         with tempfile.TemporaryDirectory() as tmpdirname:
             cli_result = self.build_docs_setup(
                 testproject_path="docs",
                 mkdocs_yml_filepath=Path(
-                    "tests/fixtures/mkdocs_lang_with_territory.yml"
+                    "tests/fixtures/mkdocs_locale_with_territory.yml"
                 ),
                 output_path=tmpdirname,
                 strict=True,
@@ -355,6 +355,58 @@ class TestBuildRss(BaseTest):
             # updated items
             feed_parsed = feedparser.parse(Path(tmpdirname) / "feed_rss_updated.xml")
             self.assertEqual(feed_parsed.feed.get("language"), "en-US")
+
+    def test_simple_build_locale_without_territory(self):
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            cli_result = self.build_docs_setup(
+                testproject_path="docs",
+                mkdocs_yml_filepath=Path(
+                    "tests/fixtures/mkdocs_locale_without_territory.yml"
+                ),
+                output_path=tmpdirname,
+                strict=True,
+            )
+
+            if cli_result.exception is not None:
+                e = cli_result.exception
+                logger.debug(format_exception(type(e), e, e.__traceback__))
+
+            self.assertEqual(cli_result.exit_code, 0)
+            self.assertIsNone(cli_result.exception)
+
+            # created items
+            feed_parsed = feedparser.parse(Path(tmpdirname) / "feed_rss_created.xml")
+            self.assertEqual(feed_parsed.feed.get("language"), "fr")
+
+            # updated items
+            feed_parsed = feedparser.parse(Path(tmpdirname) / "feed_rss_updated.xml")
+            self.assertEqual(feed_parsed.feed.get("language"), "fr")
+
+    def test_simple_build_language_specific_material(self):
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            cli_result = self.build_docs_setup(
+                testproject_path="docs",
+                mkdocs_yml_filepath=Path(
+                    "tests/fixtures/mkdocs_language_specific_material.yml"
+                ),
+                output_path=tmpdirname,
+                strict=True,
+            )
+
+            if cli_result.exception is not None:
+                e = cli_result.exception
+                logger.debug(format_exception(type(e), e, e.__traceback__))
+
+            self.assertEqual(cli_result.exit_code, 0)
+            self.assertIsNone(cli_result.exception)
+
+            # created items
+            feed_parsed = feedparser.parse(Path(tmpdirname) / "feed_rss_created.xml")
+            self.assertEqual(feed_parsed.feed.get("language"), "fr")
+
+            # updated items
+            feed_parsed = feedparser.parse(Path(tmpdirname) / "feed_rss_updated.xml")
+            self.assertEqual(feed_parsed.feed.get("language"), "fr")
 
     def test_simple_build_pretty_print_enabled(self):
         with tempfile.TemporaryDirectory() as tmpdirname:
@@ -457,6 +509,35 @@ class TestBuildRss(BaseTest):
             # cli should returns an error code (1)
             self.assertEqual(cli_result.exit_code, 1)
             self.assertIsNotNone(cli_result.exception)
+
+    def test_date(self):
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            cli_result = self.build_docs_setup(
+                testproject_path="docs",
+                mkdocs_yml_filepath=Path(
+                    "tests/fixtures/mkdocs_dates_overridden_in_dot_key.yml"
+                ),
+                output_path=tmpdirname,
+                strict=True,
+            )
+            self.assertEqual(cli_result.exit_code, 0)
+            self.assertIsNone(cli_result.exception)
+
+            feed_rss_created = feedparser.parse(
+                Path(tmpdirname) / "feed_rss_created.xml"
+            )
+            for page in feed_rss_created.entries:
+                if page.title == "Page with meta date in dot key":
+                    self.assertEqual(page.published, "Sat, 07 Oct 2023 10:20:00 +0000")
+                    break
+
+            feed_rss_updated = feedparser.parse(
+                Path(tmpdirname) / "feed_rss_updated.xml"
+            )
+            for page in feed_rss_updated.entries:
+                if page.title == "Page with meta date in dot key":
+                    self.assertEqual(page.published, "Sun, 08 Oct 2023 10:20:00 +0000")
+                    break
 
     def test_bad_date_format(self):
         # add a new page without tracking it